backend/requirements.txt
fastapi
uvicorn
langchain==0.0.316
llama-index==0.9.48
openai
openai
pydantic
python-dotenv
pdfminer.six
scikit-learn